Rename `cms.feedback` -> `cms.forms` module
