+++ metrics/api/settings/default.py
@@ -59,7 +59,7 @@ if config.APP_MODE != "INGESTION":
         "cms.whats_new",
         "cms.metrics_documentation",
         "cms.snippets",
-        "cms.feedback",
+        "cms.forms",
         "wagtail.api.v2",
         "wagtail.contrib.forms",
         "wagtail.contrib.redirects",
